requirements.txt
tkcalendar==1.6.1
mysql_connector_repackaged==0.3.1
Pillow==8.0.0
python-dotenv==0.14.0
